requirements.txt
pymongo
fastapi
uvicorn
cachetools
python-dotenv
pydantic[email]
python-multipart
apscheduler
aioredis
dependency_injector